IntakeCMD split up, Safety features added; VictorSPX now in Climb,
IntakeCMD split up into IntakeLiftCMD and IntakeWheelCMD. ClimbSUB now running on VictorPSX and not SparkMax, All motors now stop when bot disabled.

+++ Robot2020-master/src/main/java/frc/robot/Constants.java
@@ -81,7 +81,7 @@ public final class Constants {
 		public static final double BALL_AREA = Math.sqrt(0.89) * Math.PI; //m²
 		public static final double BALL_DRAG = 0.47;
 
-		public static final double MIN_DRIVEBASE_SPEED = 500;
+		public static final int MIN_DRIVEBASE_SPEED = 500;
 
 		public static final double SHOOTING_RANGE = 6.7; //m
 	}
